fix: CI for Ruby 3.5
Ruby 3.5 has added a new method `instance_variables_to_inspect` that is called when the object is inspected to define the behavior of the inspect. Because we respond to this in method_missing we try to load a configuration. This changes the respond_to_missing? method to filter this out

Co-authored-by: Maximilian Babescu <maximilian1991@gmail.com>